Cleaned kernel and fixed sgn function

+++ include/math.h
@@ -5,6 +5,6 @@
 
 int pwr(int base, int n);
 int abs(int n);
-uint8_t sgn(int n);
+int sgn(int n);
 
 #endif
